Proper usage of DB connection for a several routes.
The flask using multithreading for page rendering, so it can create troubles with sqlalchemy when a global `session` variable is used in a function, but a variable is created in one thread at start up, but used in other thread at the moment of routing.
To avoid this issue the global `session` variable has to be provided as function argument to function/methods that will be using DB connection.